Created notebook for exploratory data analysis of IQ data files (#3 & #4)
Right now it will step through a nominal and drone-present IQ captures
of the user's choice and present time and freq domain plots. It is designed
with the intention of being dynamic, it captures the metadata information
from the corresponding .txt files for the captures and uses the parameters
for proper plotting and labelling. It also allows the user to step through
the data live in 'buffer size' increments.